+++ src/test/php/io/streams/compress/unittest/Bzip2InputStreamTest.class.php
@@ -15,11 +15,6 @@ class Bzip2InputStreamTest extends DecompressingInputStreamTest {
 
   /** Erroneous data */
   protected function erroneous() {
-
-    // PHP 7.4RC1 is the first version to handle reading errors correctly, see
-    // https://github.com/php/php-src/commit/d59aac58b3e7da7ad01a194fe9840d89725ea229
-    if (PHP_VERSION_ID >= 70400) {
-      yield ["BZh61AY&SY\331<Plain data>"];
-    }
+    yield ["BZh61AY&SY\331<Plain data>"];
   }
 }